# docs/swagger.yaml
basePath: /panel/api
definitions:
...
  entity.Msg:
    properties:
      msg:
        description: Response message text
        type: string
      obj:
        description: Optional data object
      success:
        description: Indicates if the operation was successful
        type: boolean
    type: object
...
  service.ProcessState:
    enum:
    - running
    - stop
    - error
    type: string
    x-enum-comments:
      Error: Process is in error state
      Running: Process is running normally
      Stop: Process is stopped
    x-enum-descriptions:
    - Process is running normally
    - Process is stopped
    - Process is in error state
    x-enum-varnames:
    - Running
    - Stop
    - Error
  service.Status:
    properties:
      appStats:
        properties:
          mem:
            type: integer
          threads:
            type: integer
          uptime:
            type: integer
        type: object
      cpu:
        type: number
      cpuCores:
        type: integer
      cpuSpeedMhz:
        type: number
      disk:
        properties:
          current:
            type: integer
          total:
            type: integer
        type: object
      loads:
        items:
          type: number
        type: array
      logicalPro:
        type: integer
      mem:
        properties:
          current:
            type: integer
          total:
            type: integer
        type: object
      netIO:
        properties:
          down:
            type: integer
          up:
            type: integer
        type: object
      netTraffic:
        properties:
          recv:
            type: integer
          sent:
            type: integer
        type: object
      publicIP:
        properties:
          ipv4:
            type: string
          ipv6:
            type: string
        type: object
      swap:
        properties:
          current:
            type: integer
          total:
            type: integer
        type: object
      tcpCount:
        type: integer
      udpCount:
        type: integer
      uptime:
        type: integer
      xray:
        properties:
          errorMsg:
            type: string
          state:
            $ref: '#/definitions/service.ProcessState'
          version:
            type: string
        type: object
    type: object
...
host: localhost:2053
info:
  contact:
    name: API Support
    url: https://github.com/ByteProvider/3x-ui/issues
  description: API documentation for 3x-ui panel - Xray management system
  license:
    name: GPL-3.0
    url: https://github.com/ByteProvider/3x-ui/blob/main/LICENSE
  termsOfService: https://github.com/ByteProvider/3x-ui
  title: 3x-ui Panel API
  version: "2.0"
paths:
...
  /server/cpuHistory/{bucket}:
    get:
      consumes:
      - application/json
      description: Get aggregated CPU usage history for the specified time bucket
      parameters:
      - description: Time bucket in seconds (2, 30, 60, 120, 180)
        in: path
        name: bucket
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Get CPU usage history
      tags:
      - server
  /server/getConfigJson:
    get:
      consumes:
      - application/json
      description: Retrieve the Xray configuration as JSON
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Get Xray config
      tags:
      - server
  /server/getDb:
    get:
      consumes:
      - application/json
      description: Download the database file
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Download database
      tags:
      - server
  /server/getNewEchCert:
    post:
      consumes:
      - application/json
      description: Generate a new ECH certificate for the given SNI
      parameters:
      - description: SNI (Server Name Indication)
        in: formData
        name: sni
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Generate ECH certificate
      tags:
      - server
  /server/getNewUUID:
    get:
      consumes:
      - application/json
      description: Generate a new UUID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Generate UUID
      tags:
      - server
  /server/getNewVlessEnc:
    get:
      consumes:
      - application/json
      description: Generate a new VLESS encryption key
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Generate VLESS encryption key
      tags:
      - server
  /server/getNewX25519Cert:
    get:
      consumes:
      - application/json
      description: Generate a new X25519 certificate
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Generate X25519 certificate
      tags:
      - server
  /server/getNewmldsa65:
    get:
      consumes:
      - application/json
      description: Generate a new ML-DSA-65 key
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Generate ML-DSA-65 key
      tags:
      - server
  /server/getNewmlkem768:
    get:
      consumes:
      - application/json
      description: Generate a new ML-KEM-768 key
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Generate ML-KEM-768 key
      tags:
      - server
...
  /server/importDB:
    post:
      consumes:
      - multipart/form-data
      description: Import a database file and restart the Xray service
      parameters:
      - description: Database file
        in: formData
        name: db
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Import database
      tags:
      - server
  /server/installXray/{version}:
    post:
      consumes:
      - application/json
      description: Install or update Xray to the specified version
      parameters:
      - description: Xray version
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Install Xray version
      tags:
      - server
  /server/logs/{count}:
    post:
      consumes:
      - application/json
      description: Retrieve the application logs based on count, level, and syslog
        filters
      parameters:
      - description: Number of log entries
        in: path
        name: count
        required: true
        type: integer
      - description: Log level filter
        in: formData
        name: level
        type: string
      - description: Syslog filter
        in: formData
        name: syslog
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Get application logs
      tags:
      - server
  /server/restartXrayService:
    post:
      consumes:
      - application/json
      description: Restart the Xray service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Restart Xray service
      tags:
      - server
  /server/status:
    get:
      consumes:
      - application/json
      description: Get current server status including CPU, memory, disk usage and
        Xray status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entity.Msg'
            - properties:
                obj:
                  $ref: '#/definitions/service.Status'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Get server status
      tags:
      - server
  /server/stopXrayService:
    post:
      consumes:
      - application/json
      description: Stop the Xray service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Stop Xray service
      tags:
      - server
  /server/updateGeofile/{fileName}:
    post:
      consumes:
      - application/json
      description: Update the specified geo file for Xray
      parameters:
      - description: Geo file name
        in: path
        name: fileName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Update geo file
      tags:
      - server
  /server/xraylogs/{count}:
    post:
      consumes:
      - application/json
      description: Retrieve Xray logs with filtering options for direct, blocked,
        and proxy traffic
      parameters:
      - description: Number of log entries
        in: path
        name: count
        required: true
        type: integer
      - description: Log filter
        in: formData
        name: filter
        type: string
      - description: Show direct traffic
        in: formData
        name: showDirect
        type: string
      - description: Show blocked traffic
        in: formData
        name: showBlocked
        type: string
      - description: Show proxy traffic
        in: formData
        name: showProxy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Msg'
      security:
      - ApiKeyAuth: []
      summary: Get Xray logs
      tags:
      - server
...
securityDefinitions:
  ApiKeyAuth:
    description: API Key authentication. Use X-API-Key header or Authorization Bearer
      token
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"